AG-12847 - remove column event dispatcher
